initial working parallelism
right now the sdf file is read serially, then once a given batch size is
read in the [make rdkit mol -> generate fingerprint] task is done in
parallel. For running the model, all fingerprints are combined and one
process runs the model on them.

the best approach is probably to have one thread keep reading the input
file and filling a queue, a bunch of workers grabbing from the queue
when they can, and then have a final thread look for the right batch
size of fingerprints to run the model.

still looking for a better way to read the file though.